@@ -14,14 +14,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.1.2",
-    "@angular/common": "^17.1.2",
-    "@angular/compiler": "^17.1.2",
-    "@angular/core": "^17.1.2",
-    "@angular/forms": "^17.1.2",
-    "@angular/platform-browser": "^17.1.2",
-    "@angular/platform-browser-dynamic": "^17.1.2",
-    "@angular/router": "^17.1.2",
+    "@angular/animations": "^18.1.1",
+    "@angular/common": "^18.1.1",
+    "@angular/compiler": "^18.1.1",
+    "@angular/core": "^18.1.1",
+    "@angular/forms": "^18.1.1",
+    "@angular/platform-browser": "^18.1.1",
+    "@angular/platform-browser-dynamic": "^18.1.1",
+    "@angular/router": "^18.1.1",
     "body-parser": "^1.20.1",
     "crypto-js": "^4.1.1",
     "express": "^4.19.2",
@@ -31,9 +31,9 @@
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.1.2",
-    "@angular/cli": "~17.1.2",
-    "@angular/compiler-cli": "^17.1.2",
+    "@angular-devkit/build-angular": "^18.1.1",
+    "@angular/cli": "~18.1.1",
+    "@angular/compiler-cli": "^18.1.1",
     "@types/crypto-js": "^4.1.1",
     "@types/jasmine": "~4.0.0",
     "eslint-config-prettier": "^8.5.0",
@@ -46,6 +46,6 @@
     "karma-jasmine-html-reporter": "~2.0.0",
     "nodemon": "^3.0.1",
     "prettier": "^2.8.1",
-    "typescript": "~5.3.3"
+    "typescript": "~5.5.3"
   }
 }
